reorganizado el api y eliminado la clase camara
