+++ api.test2.py
@@ -1,4 +1,4 @@
-from src.api import ServerAPI
+from src.packages.api import ServerAPI
 import datetime
 from freezegun import freeze_time
 import os
