+++ src/packages/server/__init__.py
@@ -1,3 +1,2 @@
-from .camera import Camera
 from .user import User
 from .server import Server
