+++ src/packages/server/server.py
@@ -1,5 +1,4 @@
 from .user import User
-from .camera import Camera
 import json
 import os
 import random
